Added some things for better code smell

+++ backend-service/src/main/java/co/teamsphere/api/services/impl/MessageServiceImpl.java
@@ -40,7 +40,6 @@ public class MessageServiceImpl implements MessageService {
     @Override
     @Transactional
     public Messages sendMessage(SendMessageRequest req) throws UserException, ChatException {
-
         log.info("Attempting to send a message");
 
         try {
@@ -49,6 +48,11 @@ public class MessageServiceImpl implements MessageService {
 
             Chat chat = chatService.findChatById(req.getChatId());
             log.info("Found chat for sending message: {}", chat);
+
+            if (chat.getUsers().stream().noneMatch(u -> u.getId().equals(user.getId()))) {
+                log.error("User {} is not part of chat {}", user.getId(), chat.getId());
+                throw new UserException("User is not part of the chat");
+            }
 
             Messages messages = Messages.builder()
                 .chat(chat)
@@ -72,16 +76,21 @@ public class MessageServiceImpl implements MessageService {
 
     @Override
     @Transactional
-    public void deleteMessage(UUID messageId) throws MessageException {
+    public void deleteMessage(UUID messageId, UUID reqUserId) throws MessageException {
         log.info("Attempting to delete message with ID: {}", messageId);
 
         try {
-            Messages messages = findMessageById(messageId);
+            Messages messages = messageRepo.findById(messageId).orElseThrow(() -> new MessageException("Message not found with ID: " + messageId));
             log.info("Found message for deletion: {}", messages);
 
-            messageRepo.deleteById(messages.getId());
-            log.info("Message deleted successfully");
+            if (!messages.getUsername().getId().equals(reqUserId) || (messages.getChat().getIsGroup() && messages.getChat().getAdmins().stream().noneMatch(u -> u.getId().equals(reqUserId)))) {
+                log.error("User {} is not part of the chat", reqUserId);
+                throw new MessageException("User is not part of the chat");
+            }
 
+            messageRepo.deleteById(messages.getId());
+
+            log.info("Message deleted successfully");
         } catch (MessageException e) {
             log.error("Error deleting message: {}", e.getMessage());
             throw e;
@@ -93,12 +102,16 @@ public class MessageServiceImpl implements MessageService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<Messages> getChatsMessages(UUID chatId) throws ChatException {
+    public List<Messages> getChatsMessages(UUID chatId, UUID reqUserId) throws ChatException {
         log.info("Attempting to retrieve messages for chat with ID: {}", chatId);
-
         try {
             Chat chat = chatService.findChatById(chatId);
             log.info("Found chat for retrieving messages: {}", chat);
+
+            if (chat.getUsers().stream().noneMatch(u -> u.getId().equals(reqUserId))) {
+                log.error("User {} is not part of chat {}", reqUserId, chatId);
+                throw new ChatException("User is not part of the chat");
+            }
 
             List<Messages> messages = messageRepo.findMessageByChatId(chatId);
             log.info("Retrieved {} messages for chat with ID: {}", messages.size(), chatId);
